refactor: Provider 컴포넌트를 index.ts에 묶어두도록 변경
