+++ src/app/(main)/grouping/[groupId]/members/page.tsx
@@ -1,7 +1,7 @@
 import MemeberList from './components/MemberList.client';
 import MembersHeader from './components/MembersHeader';
 import { Keys, getGroupMembers } from '@/apis/groups';
-import { HydrationProvider } from '@/components/common/Provider/HydrationProvider';
+import { HydrationProvider } from '@/components/common/Provider';
 import { Suspense } from 'react';
 
 interface GroupingMembersPageProps {
