+++ src/app/(main)/grouping/components/GroupingCardList.client.tsx
@@ -1,15 +1,16 @@
 'use client';
 import { useGetGroups } from '@/apis/groups';
 import { GroupingCard } from '@/components/Card';
+import { Flex } from '@/components/Layout';
 
 export default function GroupingCardList() {
   const { data } = useGetGroups();
 
   return (
-    <section className="flex flex-col gap-8">
+    <Flex direction="column" className="gap-8">
       {data.map((groupingData) => (
         <GroupingCard groupingData={groupingData} key={groupingData.groupId} />
       ))}
-    </section>
+    </Flex>
   );
 }
