+++ src/app/(main)/profile/praise/page.tsx
@@ -1,7 +1,7 @@
 import PraiseHeader from './components/PraiseHeader';
 import ProfilePraiseDetail from './components/ProfilePraiseDetail.client';
 import { Keys, getPraises } from '@/apis/profile';
-import { HydrationProvider } from '@/components/common/Provider/HydrationProvider';
+import { HydrationProvider } from '@/components/common/Provider';
 import { QueryAsyncBoundary } from '@suspensive/react-query';
 
 export default function PraisePage() {
