+++ src/app/(main)/profile/setting/edit/page.tsx
@@ -1,7 +1,7 @@
 import ProfileEdit from './components/ProfileEdit.client';
 import { Keys, getProfile } from '@/apis/profile';
 import { RejectedFallback } from '@/components/common/ErrorBoundary';
-import { HydrationProvider } from '@/components/common/Provider/HydrationProvider';
+import { HydrationProvider } from '@/components/common/Provider';
 import { QueryAsyncBoundary } from '@suspensive/react-query';
 
 export default function page() {
